prganize the code a bit
add header files and comments
