+++ NalUnitLossSimulator/NalUnitLossSimulator/Debug/NalUnitL.DA8C8B97.tlog/NalUnitLossSimulator.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
-Debug|Win32|C:\Dev\CppSeries\NalUnitLossSimulator\|
+Debug|Win32|C:\Dev\GitHub_Clone\GitHub\LossTest\NalUnitLossSimulator\|
