+++ NalUnitLossSimulator/NalUnitLossSimulator/Release/NalUnitL.DA8C8B97.tlog/NalUnitLossSimulator.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
+Release|Win32|C:\Dev\GitHub_Clone\GitHub\LossTest\NalUnitLossSimulator\|
